backend/requirements.txt
Django==1.11.17
psycopg2==2.7.7
djangorestframework==3.9.1
flake8==3.7.5
django-cors-headers==2.4.0
djangorestframework_jwt==1.11.0
